fix(effects-docs): no crash on back from example
Hitting browser back somehow initiates javascript execution
on the /examples page (probably because gatsby has SPA like characteristics),
which was console.warn(...)ing. This triggered a log in
currently loaded /examples/:example-name page, because they are the
same virtual page (not literal page though).

This activated the monkey patched log-interception logic, which
triggered a setState call to update the virtual-console, !!WHICH!! is
all fine and good. !!BUT!! componentWillUnmount hadn't fired yet (grrrr) and
the react-internals know we were unmounting, which then causes react to
log/warn about the invalidity of calling a log method during unmount, which then causes ...another setState, which ...INFINITE LOOP.

Bummer.